+++ test/Neuroglia.UnitTests/Cases/Data/Repositories/FakeTests.cs
@@ -13,13 +13,27 @@ namespace Neuroglia.UnitTests.Cases.Data.Repositories
 {
 
     [TestCaseOrderer("Xunit.PriorityTestCaseOrderer", "Neuroglia.Xunit")]
-    public sealed class FakeTests
+    public sealed class FakeTests :
+        IDisposable
     {
+
+        public FakeTests()
+        {
+            ServiceCollection services = new();
+            services.AddLogging();
+            this.ServiceScope = services.BuildServiceProvider().CreateScope();
+        }
+        IServiceScope ServiceScope { get; }
 
         [Fact, Priority(1)]
         public void Test()
         {
             Assert.True(true);
+        }
+
+        public void Dispose()
+        {
+            this.ServiceScope.Dispose();
         }
 
     }
